requirements.txt
langchain==0.2.11
langchain-openai==0.1.17
langgraph==0.1.15
streamlit==1.37.0
pypdf==4.3.1
python-dotenv==1.0.1